LibWeb: Split LengthStyleValue out of StyleValue.{h,cpp}
